+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - (potentially) fixed rare crash when changing rooms
 - Support Commet bios.
 - Raise bio limit to 1024 characters.
+- Account switcher toggle in config.json.
 
 ## Version 1.0.0